maps: determine region type and load address
The region type can be one of the following:
misc  (default),
code  (belongs to library or executable),
exe   (belongs to the executable),
heap  (is the heap region),
stack (is the stack region)

The types misc, heap and stack are single regions which can be
determined easily. But code and exe are special. In the executable
or when loading a library, there are often three or four
consecutive regions belonging to it:
0: read,exec; 1: read; 2: read,write; 3: read,write (no file name)
0: .text; 1: .rodata; 2: .data; 3: .bss

Scanmem only cares for regions 2 and 3. But with address space
layout randomization (ASLR) and PIC/PIE the load address of the
code (start of region 0) is randomized. Subtracting the load
address from a found memory address within a code region can show
us the static memory address within the binary. This is why the
load address is important and should be determined. For other types
the region start is used.

Further commits using the region type and the load address follow.