handlers: show match region info with 'list' command
It makes things much easier if the 'list' command not only shows
the match info but the associated region info as well. Display the
region ID and the region type to get a feeling in which kind of
memory region a match is located. Subtracting the load address
of the executable or library from the match address allows us
to calculate a match offset which is the variable address within
the binary. This helps bypassing address space layout randomization
(ASLR). So display it.

The GUI needs to be changed in following commits to recognize the
new format.