handlers: improve pointer output
The pointer format should not only be used for the 'list' command
but e.g. for the 'lregions' command as well. Also '%20p' is too
long as, on x86_64 at least, there are only 6 bytes used for the
memory address. We can also get rid of the '0x' to shorten it even
further. So change it to '%12lx' for 64 bit and '%8lx' on 32 bit.

We've also noticed that there is no 'ULONGMAX' define. The correct
name would be 'ULONG_MAX'. So rename it and add a compiler warning
in case it's not defined.