ptrace: ARM: copy data for matching instead of dereferencing
On an armv7l system we've noticed an alignment issue. The kernel
detected that in arch/arm/mm/alignment.c:
Alignment trap: scanmem (7783) PC=0x3aaea6b4 Instr=0xedd00b00
Address=0x0001d191 FSR 0x001

We can't dereference a memory address which is not sizeof(void *)
aligned. So copy the data instead of dereferencing. Do this only
for ARM as it might be slower on other systems.