maps: fix and document the code regions detection
The assuption that there is always a .rodata section and .data is
consecutive to the previous region is wrong. On an armv7l system
we've found a binary without one. Furthermore, we've found a binary
on that system which has the heap region consecutive to .data of
the executable. So it has been assigned to it by mistake. With this
information it is clear now that the only way to be sure here is
to compare the filename for each possible code region. Only for the
.bss section it is still safe to assume that its filename is empty
and that its consecutive to the .data section.

Add a source code documentation for the algorithm so that others
can follow this approach.