autotools: Add --enable-libless to autoconf/automake
Add the --enable-libless configuration option to autoconf and
compile libreplacements.c if it is set. Move all AC_CHECK_LIB
macros to the case where this option is disabled.

If the option is set, set both HAVE_LIBTERMCAP and HAVE_LIBREADLINE
to 0. If it isn't, then check for the libraries termcap and
readline. This handles the case correctly where --enable-libless is
set although the libraries would be available as well.

Signed-off-by: Jonathan Pelletier <funmungus@gmail.com>
[spars: final rework fixing line length, doc and description]
Signed-off-by: Sebastian Parschauer <s.parschauer@gmx.de>