scanmem: add strdupa() and strndupa() compatibility macros
The macros strdupa() and strndupa() are GNU specific and not
available on other platforms like e.g. the BSD derivatives this way.
The build fails there. We use the GPL and the macros in the glibc
string.h are under LGPL.

So introduce adapted copies of them and use them if they aren't
defined already. This is legal as the GPL license is compatible
and stricter than the LGPL.

Reported-by: Jonathan Pelletier <funmungus@gmail.com>