gui: convert 'line' from bytes to unicode in update_scan_result
With Python 3 the memory search for strings and byte arrays doesn't
work properly and GameConqueror freezes with a Python debug message.
The scanmem output is returned line by line in bytes instead of
unicode. Python 3 intentionally doesn't convert. So convert the
string explicitly to unicode.

This has been tested with Python 2.7 and 3.4.

Reported-by: calculuswhiz