gui: convert 'bytes' from bytes to unicode in bytes2value
With Python 3 the memory search for strings and byte arrays doesn't
work properly and GameConqueror freezes with a Python debug message.
The string/bytearray value returned by scanmem is returned in bytes
instead of unicode. Python 3 intentionally doesn't convert.
So convert the string/bytearray explicitly to unicode. Also convert
the string value to str afterwards for not breaking Python 2 support.

This has been tested with Python 2.7 and 3.4.