ptrace: Optimize scan progress update for user
Currently, the scan progress update for the user is using an
expensive computation at each and every scanned byte. This is very
inefficient as it is only used to display 10 dots while processing
a memory region or all already found matches.

So restructure the code to reduce the cost of this computation. The
idea is to rarely calculate the number of scanned bytes at the next
dot and to frequently just check if this position is reached. This
produces approximately a 30% performance improvement and should have
no functional impact.

[sriemer: added handling of rounding, got rid of magic numbers]