ptrace: Optimize scan progress update for front-end
Currently, the scan progress update for the front-end is using an
expensive computation at each and every scanned byte. This is very
inefficient as it is only used to display a progress bar in the
front-end which doesn't require frequent updates.

So restructure the code to reduce the cost of this computation. The
idea is to rarely calculate the number of scanned bytes at the next
sample and to frequently just check if this position is reached.
100 samples are enough to get a resolution of 1% per sample in the
progress bar. For the initial scan calculate a progress per dot
while there are 10 dots per region to get rid of 'bytes_scanned'
handling. Also introduce an inline function for the printing of a
dot as it is very inefficient to call show_scan_progress() and to
prepare its unused parameters.

This produces approximately further 9% performance improvement on
Intel Core i5 and around 30% on AMD Phenom II CPUs. No big functional
impact is expected.

Together with the optimization for the progress for the user around
40% performance improvement is achieved on Intel Core i5 in total if
only few matches are found. With the AMD Phenom II it's around 50%.