gui: Update the scan progress only every 100 ms
The progress_watcher() routine is called via GObject.idle_add() and
is called too frequently on idle systems this way. So replace it
with GObject.timeout_add(), set the interval to 100 ms and set the
priority to GObject.PRIORITY_DEFAULT_IDLE again.

On idle systems this results in approximately 40% performance
improvement.