value: rework error handling in valtostr()
When searching for numbers with the differential search (?, <, >, =),
the GUI freezes at the point when it should list the scan results
with Python 2 and 3. Analysis has shown that it happens if "unknown"
or a number as big as the buffer is printed without a type. The
printing to the buffer is completely unchecked. All this has been
touched by commit 1f34f3e7 ("implemented dataarray representation")
the last time back in 2010.

So check if the printing to the buffer results in valid output and
make sure that there is always a type printed. For the error case use
"unknown, [unknown]" to print a type as well. Move the error handling
completely into valtostr() and change the return type to void as it
is the same for all call locations.

Reported-by: Geoff Maciolek
Fixes: GitHub issue #109