gui: hexview: add missing int conversion in AsciiText::render
There are three render() methods: OffsetText::render(),
AsciiText::render() and HexText::render(). In two of them the
division result for 'tot_lines' is correctly converted to int but
not in AsciiText::render(). During development Python 3 complained
about a float when expecting an int. So convert to int there as
well.