gui: misc: introduce unicode conversion function u()
With Python 3 everything coming from the system or the backend has
to be converted from bytes to unicode. One possible solution is to
convert to utf-8 unicode by the decode() method but that doesn't
work for random binary data read from memory or a binary file.

So introduce the unicode conversion function u() which converts
bytes to unicode for Python 3 in the same way like Python 2 does.

Reference:
http://python3porting.com/noconv.html#more-bytes-strings-and-unicode