gui: convert payload data for hexview to unicode
With Python 3 the hexview doesn't open and that crashes the GUI.
The payload data read from memory is stored in bytes instead of
unicode. So it has to be converted but converting to utf-8 with
decode() doesn't work here.
With introduction of the function u() it is possible to convert to
unicode just like Python 2 does. Hence use it here.

Fixes: GitHub issue #103