gui: hexview: correct moving the cursor in HexText
The method HexText::__on_move_cursor() doesn't handle the cursor
movement correctly: The cursor can't be moved backwards and when
moving a line down, then the cursor is also moved one position
forward. This makes cursor control unusable. This code is also
quite ugly.

So rework this method completely: When the cursor is at the end or
at the start or the movement is a line movement, then stay.
Otherwise, move one position forward or backward. Handle the spaces
used to separate hex digit pairs from each other correctly and use
this chance to remove an ugly 'if True' condition.

Fixes: GitHub issue #113