configure: detect a Python interpreter >= 2.7
The Python handling for the GUI is non-standard. The preferred GNU
automake way is the following:
http://www.gnu.org/software/automake/manual/html_node/Python.html

So introduce AM_PATH_PYTHON() to detect a Python interpreter. At
least version 2.7 which has been released end of 2008 should be
included in all still supported Linux distributions. So fail if
older versions are detected.