gui: autotools: use the Python automake macros
The Python handling for the GUI is non-standard. This is why the
Python cache isn't generated during the installation. It is
generated when calling gameconqueror instead and causes files to
stay after uninstalling GameConqueror. This is unclean. The
preferred GNU automake way is the following:
http://www.gnu.org/software/automake/manual/html_node/Python.html

With AM_PATH_PYTHON() the Python interpreter is known. So switch
over to 'python_PYTHON' and 'pythondir' to let automake generate
the Python cache upon installation. But GameConqueror.py needs to
become executable again. Use the 'install-data-hook' and 'chmod'
for that.

Reported-by: @anthraxx
Fixes: GitHub issue #116