gui: use u() instead of decode() for unicode conversion
Converting strings to utf-8 with the decode() method has the
problem that not all characters (like e.g. 0xfe) can be converted
to unicode. But with the u() function in misc.py which has been
introduced for the hexview we have a function which converts with
Python 3 to unicode just like Python 2 does. So use it everywhere
to convert to unicode.

This has been tested with Python 2.7 and 3.4.

Reported-by: @calculuswhiz
Fixes: GitHub issue #120