endianness: Move everything to the header
With the current implementation, the compiler can't optimize in the
hot memory scanning path causing function calls to fix_endianness()
all the time. This is also visible in the performance profile. Byte
swapping functions should be small and optimized. So make them all
inline by moving them to the header.

On a modern Intel Core i5 CPU this results in 7% less scan time if
only few matches are found in 500 MB of memory.