value: Compile valnowidth() inline
The function valnowidth() is very small and simple but it is called
in the hot memory scanning path several million times. In a .c file
this results in a real function call and quite slow execution.
So move it to the header and make it static inline.

On a modern Intel Core i5 CPU this results in 6% less scan time if
only few matches are found in 500 MB of memory.