value: Compile the truncval functions inline
The functions truncval() and truncval_to_flags() are very small and
simple but they are called in the hot memory scanning path several
million times. In a .c file this results in real function calls and
quite slow execution.
So move them to the header and make them static inline.

On a modern Intel Core i5 CPU this results in 5% less scan time if
searching for 0 and afterwards 10 in 500 MB of zeroed memory.