ptrace: Replace hot memset() calls in searchregions()
It looks like gcc is bad at optimizing memset() calls for unions.
In the hot memory scanning path, the union in data_value and the
checkflags union are initialized by memset() for every byte. It is
possible to zero these by setting the biggest element to 0. This
also turned out to be faster. So do exactly that.

On a modern Intel Core i5 CPU this results in 2% less scan time if
only few matches are found in 500 MB of memory.