value: Introduce zero_match_flags() and zero_value()
It is required to initialize the match_flags and value_t structures
quite frequently. This should be done at a central location so that
this can be optimized there. So introduce the static inline functions
zero_match_flags() and zero_value() for this.

Use memset() only in zero_match_flags() as it only zeros two bytes.
A memset() over ten bytes in zero_value() would be inefficient. So
split that into eight and two bytes.