ptrace: Use zero_match_flags() and zero_value()
It is required to initialize the match_flags and value_t structures
quite frequently. So use the optimized functions zero_match_flags()
and zero_value() for that.

This also addresses and parially reverts ae33fcde0869
("ptrace: Replace hot memset() calls in searchregions()"). The
performance gain is retained and also applied to peekdata().