targetmem: Compile the data_to_val functions inline
The function data_to_val_aux() is called very often in the hot
memory scanning path. Putting the function into a separate .c
file results in a real function call and quite slow execution.
So move it together with data_to_val() to the header and make
both static inline.

On a modern Intel Core i5 CPU this results in 3% less scan time if
searching for 0 and afterwards 10 in 500 MB of zeroed memory.