handlers: set: Zero value 'v' before using it
In the 'set' handler there are two locations where value 'v' isn't
initialized before use. This gives the compiler the possibility to
optimize resulting in wrong output. This issue becomes visible with
0e356afe ('targetmem: Compile the data_to_val functions inline')
and searching for consecutive floats in memory. The set value is
printed as long long int (signed 64-bit) but the set values might
be smaller (e.g. 32-bit float).
So zero the value 'v' at all locations before using and printing it.

Fixes: GitHub issue #134