handlers: set: Always print set values with '0x' prefix
The commit e87de1b70d24 ("clean up config.h , headers, and printf
formats") causes since 2010 that the printed output of setting all
matches at once and setting them separately is different. In the
case that all matches are set at once, the '0x' prefix is missing
for the set values. It was there before. So add it back.

Reference: GitHub issue #134
Fixes: e87de1b70d24 ("clean up config.h , headers, and printf formats")