getline: Introduce a getline() replacement
Android systems not only have no readline() support but they also
have no getline() support. What they have is the BSD specific
fgetln() function. So provide a getline() replacement function
based on fgetln() which provides the same functionality.

The difference between these functions is that fgetln() does no
memory allocations and it returns characters without the null
byte ('\0') termination. The minimum allocation is 120 bytes.

This is based on the work of JianXiong Zhou.

[@sriemer: major rework, wrote the commit description]