build_for_android.sh: Add a scripting file to automate building for Android.
A NDK standalone toolchain is required to build libreadline and ncurses,
and to link them in the final executable. If NDK_STANDALONE_TOOLCHAIN is not
defined, the script will prompt for the location of a standalone toolchain.
LIBREADLINE_DIR should be defined for the location of libreadline sources,
and NCURSES_DIR for ncurses sources.  If sources directories are not defined,
the script will prompt for these locations.
More information can be found by executing `./build_for_android.sh help`
README: Add a note to find Android build information in build_for_android.sh.

Signed-off-by: Jonathan Pelletier <funmungus@gmail.com>