android: Add an Android build script
An NDK standalone toolchain is required to build libreadline and
ncurses, and to link them in the final executable. If
NDK_STANDALONE_TOOLCHAIN is not defined, the script will prompt for
the location of a standalone toolchain. LIBREADLINE_DIR should be
defined for the location of libreadline sources, and NCURSES_DIR
for ncurses sources. If sources directories are not defined, the
script will prompt for these locations.
More information is printed by executing `./build_for_android.sh help`

[@sriemer: removed the ndk-build method]