gui: Don't open JSON files as binary files
JSON files are text files and no binary files. Opening them as
binary files causes Python 3 to fail with an error message like
this:

Traceback (most recent call last):
  File "/usr/share/gameconqueror/GameConqueror.py", line 382,
  in SaveCheat_Button_clicked_cb
    json.dump(obj, f);
  File "/usr/lib/python3.4/json/__init__.py", line 179, in dump
    fp.write(chunk)
TypeError: 'str' does not support the buffer interface

This especially affects saving and loading the cheat list. So always
open the cheat list as a text file.

This has been tested with Python 2.7 and 3.4.

Reported-by: @mastermax1988
Fixes: GitHub issue #141