automake: Build the getline() replacement
The getline() replacement is in place and conditional compilation
is placed into the files getline.[ch] directly. It would be possible
to always compile the getline() replacement file. But it might be
confusing that an empty file is compiled. So only compile getline.c
if the getline() function is not detected.

[@sriemer: wrote the commit description]