gui: Py2: Check again for 'long' instances
The commit 4abe7b317893 ("gui: Remove isinstance() calls for
classinfo 'long'") fixes the Python 3 issue but breaks GameConqueror
with Python 2 on 32-bit i686 systems. So check the Python version
first and only call isinstance() for classinfo 'long' if Python < 3
is detected.

This has been tested with Python 2.7 on i686 and 2.7 as well as 3.4
on x86_64.

Reference: GitHub issue #140
Fixes: 4abe7b317893 ("gui: Remove isinstance() calls for class long")