gui: Fix a crash with matches and region ID > 99
In the rare event of finding matches in regions with a region ID
> 99, GameConqueror crashes. The debug output points to the line
which splits the match offset from the region ID of the match.
Reserving two chars for the region ID is plain not enough and
relying on its length causes trouble. So split the match offset
from the region ID by searching for the '+' character.

This has been tested with Python 2.7 and 3.4.

Fixes: GitHub issue #142
Fixes: aedf0695b899 ("gui: add match offset and region type support")
Reported-by: @calculuswhiz