scanroutines: Add value range routines
Value range support for all numbers has been requested for quite
some time now. So start with adding the scan routines for ranges of
signed and unsigned integers as well as float numbers of all
supported widths. Use user values as arrays of two elements. The
type flags of both user values have to match the type flag of the
new value. Check if the new value is greater than or equal to the
first user value and less than or equal to the second user value.

Reference: GitHub PR #143, issue #76
Reported-by: Ian Douglas (@iandouglas)
[@sriemer: split commits, fixed the style, dropped MATCHUNSPECIFIED,
 wrote commit descriptions]