configure: Check for getline() and fgetln()
The fgetln()-based getline() replacement is in place and works
exactly the same way like getline(). It is safe to automatically
use it if getline() is missing but fgetln() is found. So do it like
this and just notify the user about this. If both are missing, then
fail with an error message. Use AC_CHECK_FUNCS() to define HAVE_*.