configure: Add a host OS detection
Android has some specifics which need to be detected and handled.
The common way to detect the host OS is to use AC_CANONICAL_HOST
and to check the $host_os variable. This also works for cross
compiling.
So use this method to detect Android and Linux. Set a variable
$android to "yes" if Android is detected to be used in further
commits and print a message which OS has been detected.

Reference: http://www.clearchain.com/blog/posts/autotools