autogen: Add intltoolize option --automake
The intltoolize is a bit too verbose and doesn't know about the
fact that automake is called. This triggers the message "You
should update your 'aclocal.m4' by running aclocal.". There is
the '--automake' option for this. From the manpage: "Work
silently and assume that automake is being used in software.".
So add this option to the intltoolize call.

Fixes: GitHub issue #151
Reported-by: @yz254316702