gui: Fix clicking at lock checkbox outside selection
When clicking at a 'lock' checkbox outside the selection in the cheat
list TreeView, then the checkboxes in the previously selected rows
are toggled instead of the checkbox in the row clicked at. This feels
very strange and is unexpected. This case is plain not handled within
the function cheatlist_toggle_lock_cb().

So enhance that function with checking if the current row is part of
the selection. Toggle the lock checkbox of the current row if it is
not part of the selection and return. Also move the actual toggling
to a separate function as it is called from two locations this way.

This has been tested with Python 2.7 and 3.4.

Fixes: GitHub issue #153