gui: Don't run gtk-update-icon-cache, inform instead
Running the command gtk-update-icon-cache after installation of
icons breaks the build for non-root installations like it is done
during packaging with a chroot environment. It is a rare corner
case that a user directly compiles and installs the first time
from source.

So drop running gtk-update-icon-cache and only inform the user that
he should run it in case the icons don't show up.

Fixes: a2b1405c209e ("gui: Update the GTK icon cache after install")
Reported-by: Levente Polyak (@anthraxx)