handlers: default: Prepare for ranges
With ranges we need two user values. The space used for a user value
is small. So always use an array of two elements for that and set
the 'val' pointer to the first element by default. It is also
required to change the match type. So introduce a variable for that
and set it to MATCHEQUALTO.
In the default handler 'ret' is set to false too frequently. The
error case should be the default. So initialize 'ret' with false.
Also introduce a helper function to parse numbers provided by the
user in the default handler. We will have to parse two of them with
ranges. Even argv[0] is used too frequently. Introduce a variable
'ustr' for it to make it clear that this is the user string.

Functionality remains exactly the same. Only stack usage will
increase a bit with ranges.

Reference: GitHub PR #143, issue #76