handlers: default: Add range support
The match type MATCHRANGE is available and the default handler is
prepared for ranges. There is not much missing any more.
Detect a range by searching for ".." in the string provided by
the user, parse the two user values and put them into the 'vals'
array. Set the match type to MATCHRANGE and that's all to provide
the range support.

Examples:
1.34..1.36    find values between 1.34 and 1.36 (float)
0x1..0x3      find values between 1 and 3 (hex parsing)
-3..-1        find values between -3 and -1 (signed int)

Note:
Given float values are rounded down to the next integer value if
the scan data type is not a float type and a match is an integer.
Use a float scan data type to prevent this.

Reference: GitHub PR #143, issue #76
Reported-by: Ian Douglas (@iandouglas)