gui: misc: Add range support
Split a command into two parts if it contains '..', evaluate the
operands and check if both numbers are valid integers for the given
data type. Raise an exception if not. This is the same like without
ranges. The part without ranges remains the same.

This is partially based on the work of @vyacheslavs.

Reference: GitHub PR #143, issue #76