po, gui: Add range syntax to help text, update translations
Add the range syntax to the help text so that users can see that the
new functionality is there and how to use it. This text is part of
translations. So update translations by running 'intltool-update'.
Translate new or missing strings with Google Translate and edit the
.po files to provide these translations for now. Actual translators
should have a look at them.

Reference: GitHub PR #143, issue #76