handlers: watch: Respect the scan data type
Currently, the watch handler doesn't respect the scan data type and
this is why e.g. watching floats results in watching integers
instead. Also watching strings and bytearrays is allowed but doesn't
make any sense as also integers are displayed.

So prevent watching bytearray or string and set the flags of the
first value read from memory to the match info flags.

Fixes: GitHub issue #136