scanmem: Call async-signal-unsafe functions from a child
The sighandler() uses async-signal-unsafe functions which is a
security risk. These functions have to be moved out of the signal
handling context.

So use a child process to solve this. This means: Create two
non-blocking pipes before forking an event handler process which
ignores the signals the sighandler() is registered to. Write the
signal number to the pipe from the sighandler() to the event handler,
let it print the message which signal killed scanmem and let it send
a confirmation through the other pipe back to the signal handler and
exit. Exit scanmem with _exit() from the sighandler() then to achieve
async-signal-safety. This also detaches scanmem from the target
process. Use select() to poll for input with a timeout of 1 second.
Also implement an orphan detection for the event handler and exit
there if the parent exits.

Reference: https://www.securecoding.cert.org/confluence/display/c/
SIG30-C.+Call+only+asynchronous-safe+functions+within+signal+handlers

Fixes: GitHub issue #131
Reported-by: Markus Elfring (@elfring)