gui: Hide the AddCheatDialog if trying to delete it
The AddCheatDialog ("Manually add an entry") breaks when closing it
with the "X" button. The dialog is only initialized once and always
meant to be hidden if it's not needed but this deletes it. Deletion
causes a crash or an empty dialog if trying to reopen it.
So hide the dialog in that case instead.

This has been tested with Python 2.7 and 3.4.
@12345ieee provided the reference for this.

Reference: http://faq.pygtk.org/index.py?req=show&file=faq10.021.htp

Fixes: GitHub issue #160
Reported-by: Andrea Stacchiotti (@12345ieee)