gameconqueror.in: Directly execute GameConqueror.py
There is no need to call 'env' to set the environment vars
because there is a dedicated PolicyKit policy (generated by
org.freedesktop.gameconqueror.policy.in.in) that sets the
"allow_gui" annotation to true.

This way the message of the polkit auth window can be customized
and translated by editing org.freedesktop.gameconqueror.policy.in.in.

The call to 'env' also causes trouble on openSUSE Leap 42.1.
So revert 9aba570a3e89 ("gui: Export DISPLAY and XAUTHORITY") from
PR #127. Also add command line parameters passthrough to the Python
script.

[@sriemer: enhanced description]