scanmem.1: Update to v0.16 state until examples
The scanmem man page reflects a very old state from 2010. So go
through it and test each and every command if its documentation is
the same as its implementation.
Update and enhance the man page to reflect the current v0.16 state
until the EXAMPLES section. Also add missing documentation for new
commands and reorder commands.