targetmem: Optimize memset() in add_element()
The problem with the memset() in add_element() is that it is called
in the hot scan path several million times if many memory locations
are matching. As the size to be zeroed is unknown, this results in
a real function call and costs too much performance. Most of the
time, 4 or 8 bytes have to be zeroed. So add these cases to a switch
case to let GCC optimize them and avoid the real function call.

On an Intel Xeon CPU E5-1620 v3 with DDR4 RAM this results in 6%
less scan time if everything matches when searching for 0 in 500 MB
of zeroed memory.

Reference: GitHub issue #138