GameConqueror.xml: Replace GtkButton:use-stock usage
When starting GameConqueror on openSUSE Leap 42.1 with GTK 3.16, then
the following message is displayed:

/usr/share/gameconqueror/GameConqueror.py:111: Warning: The property
GtkButton:use-stock is deprecated and shouldn't be used anymore.
It will be removed in a future version.
  self.builder.add_from_file(os.path.join(WORK_DIR, 'GameConqueror.xml'))

It has been deprecated since GTK 3.10. Drop that property and replace
it with a translatable label with a mnemonic to achieve the same.