GameConqueror.xml: Replace GtkImage:stock with 'icon-name'
When starting GameConqueror on openSUSE Leap 42.1 with GTK 3.16, then
the following message is displayed:

/usr/share/gameconqueror/GameConqueror.py:111: Warning: The property
GtkImage:stock is deprecated and shouldn't be used anymore. It will
be removed in a future version.
  self.builder.add_from_file(os.path.join(WORK_DIR, 'GameConqueror.xml'))

It has been deprecated since GTK 3.10. So replace it with the
'icon_name' property and use the new icon names if possible.

Reference: https://developer.gnome.org/gtk3/stable/GtkImage.html