handlers: add proper pager support
Add a function that checks if $PAGER works, and return a valid
file descriptor regardless of the result.

Unfortunately, there is no easy way to tell if execve() failed or the
command itself returned non-zero while we are fork()'d. We want to
know this so we can tell the user if $PAGER is really failing, i.e.,
if the command executes just fine, but returns non-zero (in the case
of linux's more) we can't know that without using a mechanism to
communicate between different processes. For simplicity's sake, I
decided to use a pipe to accomplish that.