gui: Fix a crash with backslashes in string search
GameConqueror hangs and prints a debug trace if searching for '\\\'
with data type "string" and finding matches for this. Furthermore,
it is a security risk to allow Python expressions. So remove the
eval() and repr() calls.

This works for Python 2, has been tested with 2.7 but Python 3 has
further issues to be addressed later on.

Reference: GitHub issue #172
[@sriemer: wrote the commit description]