scanmem.1: Update to v0.16 from examples to end
The scanmem example is so old that all scanmem output has changed
meanwhile. Even the nethack executable name has changed. A hint
regarding the next step (using a game trainer with found values)
and a note about issues with ASLR/PIC/PIE are missing. It should
also be mentioned that scanmem is tested on ARM and has Android
support.
So update and enhance the scanmem man page to the state of v0.16
to fix these issues. This makes the new man page complete.

Reference: GitHub issue #157