README: Rework and enhance
The README is quite ugly, has issues, is imprecise and a clear
structure is missing. So completely rework it based on the work of
Andrea Stacchiotti (@12345ieee) and enhance it with missing sections.

Users reported that the first install from source was pain in the
ass. So document build requirements, how to get documentation and
how to update the ld.so cache. Also link to the GameConqueror
requirements.

A further commit will move the GUI section to the top.

Reference: GitHub issue #157