TODO: Update to current state
Stuff like speedhack or rounding types won't be implemented. The
speedhack is impossible with the need for root privileges and there
is ranges support to make rounding obsolete. Scanning only parts of
a region sounds like a progress issue. These are fixed - so no need
for that. Still interesting is FreeBSD support as they use a port of
5 years old v0.13 right now.
Update TODO accordingly and link to the scanmem issues page on
GitHub.

Reference: GitHub issue #157