scanmem.1: Explain why scanmem requires root privileges
Many users are wondering why scanmem requires root privileges.
This is not on all distributions the case. Usually there is the
Yama security module controlling the ptrace() capabitlities since
Linux kernel 3.4. We can't avoid root privileges on such systems.
So document that in the BUGS section.

Also use the chance to replace "pidof nethack.tty" with "pgrep
nethack" as there are many variations how the nethack process is
called.

Reported-by: Andrea Stacchiotti (@12345ieee)