gui: Symlink icons instead of checking in twice
The openSUSE build complains about duplicate files in the package.
It provides a tool which can fix that by symlinking during
installation. This leads to why it doesn't make any sense to check
in the GameConqueror icons twice and to maintain them at two
locations. So symlink from gui/icons/*/*.png to gui/*.png.