gui: Greatly improve scanresult ListStore filling time
The bottleneck for the maximal number of entries displayed
in the scanresult ListStore is the time needed to fill it with
the results of a new scan.

So, use a faster and lower level API to do the work.
The time saved allows us to show 10k entries instead of 1k,
still in a reasonable time.

Fixes: #196