gui: Add deletion of scan matches
Let the user press 'Delete' or 'BackSpace' on the scanresult
treeview to delete the selected matches.
Also add a right-click menu option for the same purpose.

The match id is stored in the scanresult liststore at update time,
then the `delete [id]` command is sent to the scanmem backend when
requested.