gui: AddCheat: Add configurable length for string/bytearray
In the "Manually add cheat" dialog there is no possibility to set
the length of the non-fixed-width types string and bytearray. The
length is set to 0 which makes them unusable.

So add a field to enter a custom length for string/bytearray
between 1 and 1024 and disable it for fixed-width types.

There is NO backup length stored. So changing the variable type in
the cheat list will set it to the width of the new data type.

Reported-by: @korcankaraokcu
Reference: #192
[@sriemer: reworked the description, fixed whitespaces and style]