gui: Add configurable length for string/bytearray
The "Manually add cheat" dialog gained a field to enter a
custom length for non-fixed-width types.

Previously the length was set to 0, which made manually added
strings/bytearrays unusable.

The length is NOT stored, so changing the variable type in the
cheatlist will set it to the width of the new datatype.

Reported by: @korcankaraokcu
Related: #192