gui: Use absolute path to open 'libscanmem.so.1'
Debian requested that we provide a libscanmem1 package and
install it to e.g. /usr/lib/scanmem/ so that it is not in the
default library search path. Also the libscanmem.so symlink
shouldn't be provided in that package.

So let GameConqueror open libscanmem from the absolute path to
libscanmem.so.1. Get it from autoconf, so that we respect the
user's choice of $prefix and $libdir.

Also delete the unneeded DATA_DIR constant from consts.py.

[@sriemer: open libscanmem.so.1, enhanced commit description]