gui: hexview: Fix unicode conversion for Python 3
The function u() has trouble decoding certain escape sequences
like '\\n'. The string '\n' is not converted to the byte sequence
5C 6E but 0A instead. This also causes a traceback due to the
shorter length.

Python 2 and 3 are both able to transform raw memory into an
integer array and to successfully process it. So introduce and use
the str2bytes() function for that.

Reported-by: Sebastian Parschauer (@sriemer)
Fixes: 858acabb1975 ("gui: convert payload data for hexview to unicode")
Fixes #220