test: Drop the test directory
The tests in the 'test' directory are very old, not applicable any
more and nobody is using them since the original author of scanmem
Tavis Ormandy passed over maintenance to Lu Wang. Testing is limited
to manual tests for now and automatic tests need to be rewritten
from scratch. Volunteers welcome.
So drop the test directory.

Reported-by: Andrea Stacchiotti (@12345ieee)
Fixes #221