main: Rework and move the scanmem root warning
The scanmem root warning indicates insistently that root privileges
are somehow mandatory but that is not true for many distributions
besides Ubuntu and smartphones.
With some affected kernels the important memory regions are hidden
from the user and with other kernels reading from /proc/$pid/maps
is resulting in "Permission denied". The user will notice that root
privileges are required.
So change the message to encourage the user to use root privileges
only if some memory regions are missing and point to the man page.

It is also annoying that this message is printed when it shouldn't
like for --version and --help.
So move the message to right in front of issuing the `reset` command
after option parsing and printing the scanmem version.

Fixes #218