gui: Fix selected for context menu in scan/cheatlist
commit 69fd04515446509df6888f7fdfe12dc9eb9db7c9 upstream.

The context menu of the lists takes data from the entry
right-clicked last which is not the correct entry if the menu
is triggered by the menu key on keyboard.
The menu items "Browse this address" and "Scan for this address"
are affected.

So use the first one of the selected entries which is guaranteed
to be the intended one in both cases.

Reported-by: @avseredyuk
Fixes #230