license: Add lgpl-3.0.txt
A license change from GPLv3 to LGPLv3 is planned for libscanmem so
that tools using libscanmem are not forced to use GPLv3 as well.
Furthermore, this is a requirement for proper packaging. Copyright
owners approved or didn't inform us about any objections.
So add the LGPLv3 license file as a first step.

Reference: #184