scanmem: Rework version and copyright printing
The version and copyright information of scanmem is part of
libscanmem. printversion() always prints the copyright information
as well. This is very unclean and it is annoying to see the
copyright information too often.

So introduce a printversion() function pointer and move the version
and copyright printing to scanmem (main.c). Set the pointer to a new
function printlibversion() by default and include the libscanmem
version in printversion(). Don't print the copyright information
in there any more. Introduce a new function printcopyright() which
behaves like the old printversion() function instead. Only call it
when scanmem starts or "scanmem --help" is run. Also update the
copyright information.

Reference: #186